django/requirements/base.txt
django
django-address
djangorestframework
markdown
django-filter
pyyaml
django-cors-headers
django-gravatar2
raven
